+++ djangocms_frontend/contrib/carousel/forms.py
@@ -7,11 +7,13 @@ from filer.models import Image
 
 from djangocms_frontend.fields import (
     AttributesFormField,
+    ButtonGroup,
     TagTypeFormField,
     TemplateChoiceMixin,
 )
 
 from ... import settings
+from ...common.background import BackgroundFormMixin
 from ...fields import HTMLFormField
 from ...helpers import first_choice
 from ...models import FrontendUIItem
@@ -20,8 +22,8 @@ from ..link.forms import AbstractLinkForm
 from .constants import (
     CAROUSEL_ASPECT_RATIO_CHOICES,
     CAROUSEL_PAUSE_CHOICES,
-    CAROUSEL_RIDE_CHOICES,
     CAROUSEL_TEMPLATE_CHOICES,
+    CAROUSEL_TRANSITION_CHOICES,
 )
 
 mixin_factory = settings.get_forms(carousel)
@@ -46,6 +48,7 @@ class CarouselForm(mixin_factory("Carousel"), TemplateChoiceMixin, EntangledMode
                 "carousel_pause",
                 "carousel_ride",
                 "carousel_wrap",
+                "carousel_transition",
                 "attributes",
             ]
         }
@@ -93,11 +96,12 @@ class CarouselForm(mixin_factory("Carousel"), TemplateChoiceMixin, EntangledMode
             '"mouseleave". If set to "false", hovering over the carousel '
             "won't pause it."
         ),
+        widget=ButtonGroup(attrs=dict(property="text")),
     )
-    carousel_ride = forms.ChoiceField(
-        label=_("Ride"),
-        choices=CAROUSEL_RIDE_CHOICES,
-        initial=first_choice(CAROUSEL_RIDE_CHOICES),
+    carousel_ride = forms.BooleanField(
+        label=_("Auto start"),
+        initial=True,
+        required=False,
         help_text=_(
             "Autoplays the carousel after the user manually cycles the "
             'first item. If "carousel", autoplays the carousel on load.'
@@ -121,6 +125,16 @@ class CarouselForm(mixin_factory("Carousel"), TemplateChoiceMixin, EntangledMode
             "according to the selected ratio."
         ),
     )
+    carousel_transition = forms.ChoiceField(
+        label=_("Transition"),
+        choices=CAROUSEL_TRANSITION_CHOICES,
+        required=False,
+        initial=CAROUSEL_TRANSITION_CHOICES[0][0],
+        help_text=_(
+            "Determines if slides change by sliding or fading."
+        ),
+        widget=ButtonGroup(attrs=dict(property="text")),
+    )
     attributes = AttributesFormField(
         excluded_keys=[
             "id",
@@ -135,7 +149,7 @@ class CarouselForm(mixin_factory("Carousel"), TemplateChoiceMixin, EntangledMode
 
 
 class CarouselSlideForm(
-    mixin_factory("CarouselSlide"), AbstractLinkForm, EntangledModelForm
+    mixin_factory("CarouselSlide"), AbstractLinkForm, BackgroundFormMixin, EntangledModelForm
 ):
     """
     Components > "Slide" Plugin
@@ -160,6 +174,7 @@ class CarouselSlideForm(
         queryset=Image.objects.all(),
         to_field_name="id",
         label=_("Slide image"),
+        required=False,
     )
     carousel_content = HTMLFormField(
         label=_("Content"),